Fix memory leak in memory allocator
The problem was that I was passing in also the ID3D12Resource* to the D3D12MA::Allocator::CreateResource() function but in the destructor of the MemoryAllocation I was only freeing the D3D12MA::Allocation object. The problem is fixed for now by passing in nullptr to the D3D12MA::Allocator::CreateResource(), which is usually how the library is used.